pyproject.toml
[project]
name = "rue"
dynamic = ["version"]
description = "Testing Framework for AI Software"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Nikita Shirobokov", email = "shirobokov@me.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "anthropic[vertex]>=0.72.0",
    "pydantic-settings>=2.11.0",
    "httpx>=0.28.1",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-openai>=0.49.0",
    "opentelemetry-instrumentation-anthropic>=0.49.0",
    "pydantic-ai>=1.67.0",
    "typer>=0.24.1",
    "tomlkit>=0.14.0",
    "cloudpickle>=3.1.2",
    "msgspec>=0.21.1",
    "deepdiff>=9.0.0",
    "pycrdt>=0.12.50",
    "pyturso>=0.5.1",
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14"
]

[project.urls]
Repository = "https://github.com/NickSherrow/rue"

[project.entry-points.pytest11]
rue = "rue.pytest_plugin"
[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]
lint = [
    "mypy==1.19.1",
    "ruff==0.15.4",
]

[project.scripts]
rue = "rue.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/rue"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]
pythonpath = ["."]

[tool.ruff]
line-length = 80
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "B",
    "D",
    "E",
    "F",
    "I",
    "N",
    "PT",
    "RUF",
    "UP",
    "W",
    "PLC0415",
]
ignore = [
    "ANN401", # Allow Any when the boundary is intentionally dynamic
    "D107",   # Missing docstring in __init__
    "RUF012", # Conflicts with Pydantic
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D",
    "ANN", # TODO: Relax type annotations in tests
]

[tool.ruff.format]
quote-style = "double" # Like Black, use double quotes for strings
indent-style = "space" # Like Black, use spaces for indentation
docstring-code-format = true
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
enable_error_code = "deprecated"

[tool.rue.predicates.all_predicates]
model = "openai:gpt-5.4"
